+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/smartbar/quickaction/QuickAction.kt
@@ -295,6 +295,8 @@ fun QuickAction.computeDisplayName(evaluator: ComputingEvaluator): String {
             KeyCode.ARROW_DOWN -> R.string.quick_action__arrow_down
             KeyCode.ARROW_LEFT -> R.string.quick_action__arrow_left
             KeyCode.ARROW_RIGHT -> R.string.quick_action__arrow_right
+            KeyCode.PAGE_UP -> R.string.quick_action__page_up
+            KeyCode.PAGE_DOWN -> R.string.quick_action__page_down
             KeyCode.CLIPBOARD_CLEAR_PRIMARY_CLIP -> R.string.quick_action__clipboard_clear_primary_clip
             KeyCode.CLIPBOARD_COPY -> R.string.quick_action__clipboard_copy
             KeyCode.CLIPBOARD_CUT -> R.string.quick_action__clipboard_cut
@@ -339,6 +341,8 @@ fun QuickAction.computeTooltip(evaluator: ComputingEvaluator): String {
             KeyCode.ARROW_DOWN -> R.string.quick_action__arrow_down__tooltip
             KeyCode.ARROW_LEFT -> R.string.quick_action__arrow_left__tooltip
             KeyCode.ARROW_RIGHT -> R.string.quick_action__arrow_right__tooltip
+            KeyCode.PAGE_UP -> R.string.quick_action__page_up__tooltip
+            KeyCode.PAGE_DOWN -> R.string.quick_action__page_down__tooltip
             KeyCode.CLIPBOARD_CLEAR_PRIMARY_CLIP -> R.string.quick_action__clipboard_clear_primary_clip__tooltip
             KeyCode.CLIPBOARD_COPY -> R.string.quick_action__clipboard_copy__tooltip
             KeyCode.CLIPBOARD_CUT -> R.string.quick_action__clipboard_cut__tooltip